requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
python-dotenv==1.0.0
deepdiff==8.4.2
email-validator==2.1.1
eventlet==0.33.3
Werkzeug==2.3.7
